feat: docker adicionado ao projeto

+++ Quotes/Class/Price.cs
@@ -1,8 +1,8 @@
 public class Price{
-    public float Value {get; set;}
+    public Decimal Value {get; set;}
     public string Date {get; set;}
 
-    public Price(string date, float value){
+    public Price(string date, Decimal value){
         Value = value;
         Date = date;
     }
